+++ CHANGES.rst
@@ -5,5 +5,8 @@ Changelog
 1.0a1 (unreleased)
 ------------------
 
+- Add get_work_location method.
+  [fngaha]
+
 - Initial release.
   [sdelcourt]